// jovo-platforms/jovo-platform-googleassistant/src/conversational-actions/core/Interfaces.ts
export type HorizontalAlignment = 'UNSPECIFIED' | 'LEADING' | 'CENTER' | 'TRAILING';
export type ImageFill = 'UNSPECIFIED' | 'GRAY' | 'WHITE' | 'CROPPED';
// ...
export type AccountLinkingStatus = 'ACCOUNT_LINKING_STATUS_UNSPECIFIED' | 'NOT_LINKED' | 'LINKED';
export type UserVerificationStatus = 'USER_VERIFICATION_STATUS_UNSPECIFIED' | 'GUEST' | 'VERIFIED';
// ...
export interface Params {
  [key: string]: any;
}

export interface Handler {
  name?: string;
}
export interface Intent {
  name: string;
  params: Record<string, IntentParameterValue>;
  query?: string;
}

export interface IntentParameterValue {
  original: string;
  resolved: string;
}

export interface Simple {
  speech?: string;
  text?: string;
}

export interface Image {
  url: string;
  alt: string;
  height?: number;
  width?: number;
}
export interface OpenUrl {
  url: string;
  hint: UrlHint;
}

export interface Link {
  name: string;
}
export interface Card {
  title: string;
  subtitle?: string;
  text: string; // required, unless image is present
  image?: Image;
  imageFill?: ImageFill;
  button?: Link;
}

export interface TableColumn {
  header: string;
  align: HorizontalAlignment;
}

export interface TableCell {
  text: string;
}

export interface TableRow {
  cells: TableCell[];
  divider: boolean;
}

export interface Table {
  title: string;
  subtitle?: string;
  image?: Image;
  columns: TableColumn[];
  rows: TableRow[];
  button?: Link;
}
// ...
export interface User {
  locale: string;
  params: Params;
  accountLinkingStatus: AccountLinkingStatus;
  verificationStatus: UserVerificationStatus;
  lastSeenTime: string;
  engagement?: Engagement;
  packageEntitlements?: PackageEntitlement[];
}

export interface Engagement {
  pushNotificationIntents?: IntentSubscription[];
  dailyUpdateIntents?: IntentSubscription[];
}
export interface IntentSubscription {
  intent: string;
  contentTitle: string;
}

export interface PackageEntitlement {
  packageName: string;
  entitlements: Entitlement[];
}

export interface Entitlement {
  sku: string;
  skuType: SkuType;
  inAppDetails: SignedData;
}