// jovo-platforms/jovo-platform-googleassistant/src/core/google-assistant-enums.ts
export enum EnumGoogleAssistantRequestType {
  ON_PERMISSION = 'ON_PERMISSION',
  ON_SIGN_IN = 'ON_SIGN_IN',
  ON_CONFIRMATION = 'ON_CONFIRMATION',
  ON_DATETIME = 'ON_DATETIME',
  ON_PLACE = 'ON_PLACE',
  ON_HEALTH_CHECK = 'ON_HEALTH_CHECK',
  ON_NEW_SURFACE = 'ON_NEW_SURFACE',

  ON_NOINPUT1 = 'ON_NOINPUT1',
  ON_NOINPUT2 = 'ON_NOINPUT2',
  ON_NOINPUTFINAL = 'ON_NOINPUTFINAL',
  ON_NOMATCH1 = 'ON_NOMATCH1',
  ON_NOMATCH2 = 'ON_NOMATCH2',
  ON_NOMATCHFINAL = 'ON_NOMATCHFINAL',
}